refactor: file structure for pipelines
